+++ native/src/main/java/io/ballerina/lib/redis/StringCommands.java
@@ -26,6 +26,7 @@ import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
 
 import static io.ballerina.lib.redis.utils.ConversionUtils.createBError;
+import static io.ballerina.lib.redis.utils.ConversionUtils.createBNilableStringArrayFromKeyValueList;
 import static io.ballerina.lib.redis.utils.ConversionUtils.createBStringArrayFromKeyValueList;
 import static io.ballerina.lib.redis.utils.ConversionUtils.createMapFromBMap;
 import static io.ballerina.lib.redis.utils.ConversionUtils.createStringArrayFromBArray;
@@ -295,7 +296,8 @@ public class StringCommands {
     }
 
     /**
-     * Get the values of all the given keys.
+     * Get the values of all the given keys. Fails if any key has no value; use
+     * {@link #mGetNilable(BObject, BArray)} if a key might not exist.
      *
      * @param redisClient Client from the Ballerina redis client
      * @param keys        The keys of which the values need to be retrieved
@@ -305,6 +307,23 @@ public class StringCommands {
         try {
             RedisStringCommandExecutor executor = getConnection(redisClient).getStringCommandExecutor();
             return createBStringArrayFromKeyValueList(executor.mGet(createStringArrayFromBArray(keys)));
+        } catch (Throwable e) {
+            return createBError(e);
+        }
+    }
+
+    /**
+     * Get the values of all the given keys. Unlike {@link #mGet(BObject, BArray)}, a key that has
+     * no value is represented as {@code ()} in the returned array instead of failing.
+     *
+     * @param redisClient Client from the Ballerina redis client
+     * @param keys        The keys of which the values need to be retrieved
+     * @return Array of values at the specified keys, with {@code ()} for a missing key
+     */
+    public static Object mGetNilable(BObject redisClient, BArray keys) {
+        try {
+            RedisStringCommandExecutor executor = getConnection(redisClient).getStringCommandExecutor();
+            return createBNilableStringArrayFromKeyValueList(executor.mGet(createStringArrayFromBArray(keys)));
         } catch (Throwable e) {
             return createBError(e);
         }